ISSUE #5874 - move sorting context higher so that sorting is not reset when changing filters/models etc.
