+++ frontend/src/v5/ui/routes/dashboard/projects/tickets/ticketsTable/ticketsTable.component.tsx
@@ -44,7 +44,7 @@ import { ResizableTableContext, ResizableTableContextComponent } from '@controls
 import { templateAlreadyFetched } from '@/v5/store/tickets/tickets.helpers';
 import { TicketsTableContext, TicketsTableContextComponent } from './ticketsTableContext/ticketsTableContext';
 import { useContextWithCondition } from '@/v5/helpers/contextWithCondition/contextWithCondition.hooks';
-import { selectTicketsHaveBeenFetched } from '@/v5/store/tickets/tickets.selectors';
+import { selectTicketPropertyByName, selectTicketsHaveBeenFetched } from '@/v5/store/tickets/tickets.selectors';
 import { getState } from '@/v5/helpers/redux.helpers';
 import { useWatchPropertyChange } from './useWatchPropertyChange';
 import { getAvailableColumnsForTemplate } from './ticketsTableContext/ticketsTableContext.helpers';
@@ -56,8 +56,11 @@ import { FilterSelection } from '@components/viewer/cards/cardFilters/filtersSel
 import { CardFilters } from '@components/viewer/cards/cardFilters/cardFilters.component';
 import { deserializeFilter, getNonCompletedTicketFilters, getTemplateFilter, serializeFilter } from '@components/viewer/cards/cardFilters/filtersSelection/tickets/ticketFilters.helpers';
 import { useRealtimeFiltering } from './useRealtimeFiltering';
-import { isEqual } from 'lodash';
+import { isEqual, orderBy } from 'lodash';
 import { formatMessage } from '@/v5/services/intl';
+import { SortedTableComponent } from '@controls/sortedTableContext/sortedTableContext';
+import { BaseProperties, IssueProperties } from '@/v5/ui/routes/viewer/tickets/tickets.constants';
+import { getAssigneeDisplayNamesFromTicket, sortAssignees } from './ticketsTableGroupBy.helper';
 
 const paramToInputProps = (value, setter) => ({
 	value,
@@ -305,6 +308,38 @@ export const TicketsTable = ({ isNewTicketDirty, setTicketValue }: TicketsTableP
 		setTemplate(templates[0]._id);
 		return null;
 	}
+	const assigneesSort = (items: ITicket[], order) => orderBy(
+		items.map(sortAssignees),
+		[
+			(item) => getAssigneeDisplayNamesFromTicket(item).length,
+			(item) => getAssigneeDisplayNamesFromTicket(item).join(),
+		],
+		[order, order],
+	);
+
+	const sortTicketsByProperty = (items: ITicket[], order, column: string) =>  orderBy(
+		items,
+		(item) => {
+			const sortingElement = selectTicketPropertyByName(getState(), item._id, column);
+			return sortingElement?.toLowerCase?.().trim?.() ?? sortingElement;
+		},
+		order,
+	);
+
+	const ticketCodeSort = (items: ITicket[], order) => {
+		return orderBy(
+			items,
+			(item) => selectTicketPropertyByName(getState(), item._id, 'number'),
+			order,
+		);
+	};
+	const customSortingFunctions = (column: string) => {
+		if (column === 'modelName') return null; // uses the default sorting function from sortcontext
+		if (column === `properties.${IssueProperties.ASSIGNEES}` ) return assigneesSort;
+		if (column === 'id') return ticketCodeSort;
+
+		return sortTicketsByProperty;
+	};
 
 	const containers = ContainersHooksSelectors.selectContainers();
 	const federations = FederationsHooksSelectors.selectFederations();
@@ -367,7 +402,9 @@ export const TicketsTable = ({ isNewTicketDirty, setTicketValue }: TicketsTableP
 					</FlexContainer>
 				</FiltersContainer>
 				<CardFilters />
-				<TicketsTableContent tickets={filteredTickets} setTicketValue={setTicketValue} selectedTicketId={ticketId} template={selectedTemplate}/>
+				<SortedTableComponent items={tickets} sortingColumn={BaseProperties.CREATED_AT} customSortingFunctions={customSortingFunctions}>
+					<TicketsTableContent tickets={filteredTickets} setTicketValue={setTicketValue} selectedTicketId={ticketId} template={selectedTemplate}/>
+				</SortedTableComponent>
 			</TicketsTableLayout>
 		</TicketsFiltersContextComponent>
 	);
